pyproject.toml
[tool.poetry]
name = "prompt-optimizer"
version = "0.1.0"
description = "Prompt optimizasyon ve analiz sistemi"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "~3.12"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
openai = "^1.3.0"
pymongo = "4.5.0"
python-dotenv = "^1.0.0"
pydantic = "^2.4.2"
tiktoken = "^0.5.1"
motor = "3.3.1"
pydantic-settings = "^2.1.0"
jinja2 = "^3.1.3"
python-multipart = "^0.0.9"
lime = "^0.2.0"
nltk = "^3.8.1"
numpy = "^1.24.0"
scikit-learn = "^1.3.0"
matplotlib = "^3.8.0"
plotly = "^5.18.0"
pandas = "^2.1.0"
seaborn = "^0.13.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.10.1"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 